+++ aeron-driver/src/main/c/aeron_ipc_publication.c
@@ -175,12 +175,12 @@ void aeron_ipc_publication_close(aeron_counters_manager_t *counters_manager, aer
 {
     aeron_subscribable_t *subscribable = &publication->conductor_fields.subscribable;
 
-    aeron_counters_manager_free(counters_manager, (int32_t)publication->pub_lmt_position.counter_id);
-    aeron_counters_manager_free(counters_manager, (int32_t)publication->pub_pos_position.counter_id);
+    aeron_counters_manager_free(counters_manager, publication->pub_lmt_position.counter_id);
+    aeron_counters_manager_free(counters_manager, publication->pub_pos_position.counter_id);
 
     for (size_t i = 0, length = subscribable->length; i < length; i++)
     {
-        aeron_counters_manager_free(counters_manager, (int32_t)subscribable->array[i].counter_id);
+        aeron_counters_manager_free(counters_manager, subscribable->array[i].counter_id);
     }
     aeron_free(subscribable->array);
